Modify: 컴포넌트명 수정, 1차 리뷰 수정 요청 사항 반영 Add: 카테고리 타이틀 레이아웃 추가
