+++ src/Pages/ProductList/Components/Product.js
@@ -1,32 +1,27 @@
 import React, { Component } from 'react';
 
-class Item extends Component {
+class Product extends Component {
   render() {
     const { imgUrl, sex, name, sale_per, standard_price, color, stars } = this.props.product;
     const { starNumToArr } = this.props;
 
     return (
-      <div className="item_container">
-        <div className="item_block">
-          <div className="item_img">
+      <div className="Product">
+        <div className="product_block">
+          <div className="product_img">
             <img src={imgUrl} alt="상품 이미지"/>
-            <div className="item_color">{`+ ${color} ` + (color === 1 ? 'color' : 'colors')}</div>
+            <div className="product_color">{`+ ${color} ` + (color === 1 ? 'color' : 'colors')}</div>
           </div>
-          <div className="item_info">
+          <div className="product_info">
             <div className="sale_per">{`${sale_per}% 세일`}</div>
-            <div className="item_name">
+            <div className="product_name">
               <a href='#'>({sex}) {name}</a>
             </div>
             <div className="stars_container">
               <ul>
                 {
                   starNumToArr(stars).map((e, index) => {
-                    if (e === 0) {
-                      return <li key={index}><i className="fas fa-star"></i></li>;
-                    }
-                    else {
-                      return <li key={index}><i className="far fa-star"></i></li>;
-                    }
+                    return <li key={index}><i className={`${ e === 0 ? 'fas' : 'far'} fa-star`}></i></li>;
                   })
                 }
               </ul>
@@ -42,4 +37,4 @@ class Item extends Component {
   }
 }
 
-export default Item;
+export default Product;
