+++ src/Pages/ProductList/data.js
@@ -1,0 +1,155 @@
+export const PRODUCT_DATA = [
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dw4e4f26c1/AH1952-20C_166_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "브이넥 울 스웨터",
+    sale_per: 30,
+    stars: 3,
+    standard_price: 229000,
+    color: 2
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwda741ab7/AH2048-20C_HDE_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "재생 캐시미어 원단 터틀넥 스웨터",
+    sale_per: 30,
+    stars: 2,
+    standard_price: 399000,
+    color: 2
+  },
+  { // color: 1인 경우 -> + 1 color
+    // color: 2이상인 경우 -> + n colors
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dweaf8e7d7/AH2348-20C_D9A_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "라코스테 스포츠 브리더블 울 골프 스웨터",
+    sale_per: 30,
+    stars: 1,
+    standard_price: 259000,
+    color: 1
+  },
+  { // 할인율 다르게 적용해보기
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwa49b70d6/AH4832-20C_Z0W_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "베이직 사선 조직 라운드-넥 풀오버",
+    sale_per: 90,
+    stars: 4,
+    standard_price: 259000,
+    color: 3
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dw4e4f26c1/AH1952-20C_166_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "브이넥 울 스웨터",
+    sale_per: 30,
+    stars: 3,
+    standard_price: 229000,
+    color: 2
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwda741ab7/AH2048-20C_HDE_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "재생 캐시미어 원단 터틀넥 스웨터",
+    sale_per: 30,
+    stars: 2,
+    standard_price: 399000,
+    color: 2
+  },
+  { // color: 1인 경우 -> + 1 color
+    // color: 2이상인 경우 -> + n colors
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dweaf8e7d7/AH2348-20C_D9A_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "라코스테 스포츠 브리더블 울 골프 스웨터",
+    sale_per: 30,
+    stars: 1,
+    standard_price: 259000,
+    color: 1
+  },
+  { // 할인율 다르게 적용해보기
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwa49b70d6/AH4832-20C_Z0W_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "베이직 사선 조직 라운드-넥 풀오버",
+    sale_per: 90,
+    stars: 5,
+    standard_price: 259000,
+    color: 3
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dw4e4f26c1/AH1952-20C_166_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "브이넥 울 스웨터",
+    sale_per: 30,
+    stars: 3,
+    standard_price: 229000,
+    color: 2
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwda741ab7/AH2048-20C_HDE_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "재생 캐시미어 원단 터틀넥 스웨터",
+    sale_per: 30,
+    stars: 2,
+    standard_price: 399000,
+    color: 2
+  },
+  { // color: 1인 경우 -> + 1 color
+    // color: 2이상인 경우 -> + n colors
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dweaf8e7d7/AH2348-20C_D9A_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "라코스테 스포츠 브리더블 울 골프 스웨터",
+    sale_per: 30,
+    stars: 1,
+    standard_price: 259000,
+    color: 1
+  },
+  { // 할인율 다르게 적용해보기
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwa49b70d6/AH4832-20C_Z0W_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "베이직 사선 조직 라운드-넥 풀오버",
+    sale_per: 90,
+    stars: 5,
+    standard_price: 259000,
+    color: 3
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dw4e4f26c1/AH1952-20C_166_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "브이넥 울 스웨터",
+    sale_per: 30,
+    stars: 3,
+    standard_price: 229000,
+    color: 2
+  },
+  {
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwda741ab7/AH2048-20C_HDE_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "재생 캐시미어 원단 터틀넥 스웨터",
+    sale_per: 30,
+    stars: 2,
+    standard_price: 399000,
+    color: 2
+  },
+  { // color: 1인 경우 -> + 1 color
+    // color: 2이상인 경우 -> + n colors
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dweaf8e7d7/AH2348-20C_D9A_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "라코스테 스포츠 브리더블 울 골프 스웨터",
+    sale_per: 30,
+    stars: 1,
+    standard_price: 259000,
+    color: 1
+  },
+  { // 할인율 다르게 적용해보기
+    imgUrl: "https://imageapac1.lacoste.com/dw/image/v2/BBCL_PRD/on/demandware.static/-/Sites-master/default/dwa49b70d6/AH4832-20C_Z0W_20.jpg?imwidth=915&impolicy=product",
+    sex: "남성",
+    name: "베이직 사선 조직 라운드-넥 풀오버",
+    sale_per: 90,
+    stars: 5,
+    standard_price: 259000,
+    color: 3
+  }
+]
+
+// 품절인 경우
+// 평점 없는 경우
+// 할인 없는 경우
+// 다른 색 없는 경우
