+++ src/Routes.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 import Login from './Pages/Login/Login';
 import Main from './Pages/Main/Main';
-import Product from "./Pages/Product/Product";
+import ProductList from "./Pages/ProductList/ProductList";
 import ProductDetail from './Pages/ProductDetail/ProductDetail';
 
 
@@ -16,7 +16,7 @@ class Routes extends React.Component {
         <Switch>
           <Route exact path='/' component={Main} />
           <Route exact path='/login' component={Login} />
-          <Route exact path='/product' component={Product} />
+          <Route exact path='/productList' component={ProductList} />
           <Route exact path='/product_detail' component={ProductDetail} />
         </Switch>
       </Router>
